--- api/src/index.js
@@ -1,11 +1,0 @@
-export {
-  handleStripeWebhooks,
-  generateStripeMockEvent,
-  stripe,
-} from "./lib";
-
-// export schemas sdl
-export * from "./schemas";
-
-// export services
-export * from "./services";
