+++ api/src/lib/index.ts
@@ -1,5 +1,3 @@
-import { mockHttpEvent } from "@redwoodjs/testing/api";
-
 import Stripe from "stripe";
 
 export const stripe = new Stripe(process.env.STRIPE_SECRET_KEY, {
@@ -89,33 +87,4 @@ export const handleStripeWebhooks = async (
       results: unverifiedStripeEvent,
     };
   }
-};
-
-// Generates a test Stripe mock event.
-export const generateStripeMockEvent = () => {
-  const payload = JSON.stringify(
-    {
-      id: "evt_test_webhook",
-      object: "event",
-    },
-    null,
-    2,
-  );
-
-  process.env.STRIPE_WEBHOOK_SK = "whsec_test_secret";
-
-  // See https://github.com/stripe/stripe-node/blob/master/README.md#testing-webhook-signing.
-  const header = stripe.webhooks.generateTestHeaderString({
-    payload,
-    secret: process.env.STRIPE_WEBHOOK_SK,
-  });
-
-  const httpEvent = mockHttpEvent({
-    body: payload,
-    headers: {
-      "stripe-signature": header,
-    },
-  });
-
-  return httpEvent;
 };
