+++ api/src/services/subscription/subscription.ts
@@ -1,4 +1,4 @@
-import { stripe } from "../../lib";
+import { stripe } from "../../lib.js";
 
 export const listStripeSubscriptions = async ({ data: { params } }) => {
   const subscriptions = await stripe.subscriptions.list({ ...params });
